Move poke_latest function to make_request for reuse
